requirements.txt
nltk
lxml
pymongo
beautifulsoup
goose-extractor
beautifulsoup4
pattern
scipy
scikit-learn
flask
